pyproject.toml
[project]
name = "hotel-reservation-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "google-cloud-storage>=3.1.0",
    "imbalanced-learn>=0.13.0",
    "ipykernel>=6.29.5",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.3",
    "mlflow>=2.22.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
]
